feat: Implement exhaustive I/O options
From CLI args, file, or `stdin` to `stdout`, or file.

This commit prepares the next step: streaming encryption. The algorithms
already accept streams, now the CLI is ready to provide the streams. The
missing piece is connecting the two.

As of now, and as a temporary solution, the streams are fully sinked
into a string, and then the string is streamed into the ciphers and the
base64 encoder/decoder. The goal is to completely avoid the "read all
into memory" part and stream everything from input into output.